Avoid locking problem when creating PRIMARY KEY constraints
This commit removes the creation of the PRIMARY KEY constraints at the same
time as creating the indexes. The problem was that since we were creating the
indexes in parallel the ALTER TABLE ADD PRIMARY KEY statement was queued up
behind a CREATE INDEX statement on the same table. The solution is to just
add UNIQUE indexes in parallel and create the constraints later.